Fix(client): TTTK-52 Moved examples to new file
